move redis under db package
