+++ src/moderation/user.py
@@ -9,8 +9,8 @@ from quart_auth import AuthUser, Action
 from werkzeug.security import check_password_hash, generate_password_hash
 from configs import mod_conf
 from db import db_m
+from db.redis import get_redis
 from enums import DbPool
-from redis_cache import get_redis
 
 REDIS_MOD_DB: int = mod_conf.get('redis_db', 1)
